requirements.txt
discord.py
python-dotenv
discord
asyncio
requests
eth_abi